backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
asyncpg==0.29.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg[sa]==0.29.0